user-service/requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
Werkzeug==2.1.2
